requirements.txt
requests==2.31.0
openai-whisper==20231117
numpy==1.26.2
torch==2.0.0
tqdm==4.66.1
more-itertools==10.1.0
transformers==4.35.2
ffmpeg-python==0.2.0
ollama-python==0.1.6
